fix(Comments): Renamed common types and errorMiddleware files
